intermediate/poc-06-rag-multiple-documents/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and related
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# OpenAI
openai==1.6.1

# Vector store
faiss-cpu==1.7.4

# Document loaders
PyPDF2==3.0.1
python-docx==1.1.0
unstructured==0.11.6
markdown==3.5.1

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
